refactor(onboarding): float wizard, guard popover, split AppDelegate
Address review + live-test findings on the discovery beat:

- Float the onboarding window (.floating): granting Accessibility
  starts tiling, which raised other apps' windows over the wizard
  mid-flow. It always closes on finish, so nothing floats after.
- Guard the menu-bar popover on the icon being on-screen: an
  auto-hidden menu bar parks the status item offscreen, so
  anchoring there misplaced the popover to a garbage corner. Skip
  it then — the card's "get back from the menu bar" copy covers
  that user (ui-designer-confirmed the popover fires only on the
  Not Now / close routes, never alongside Open Settings).
- Fire the popover *after* the app demotes to .accessory, from the
  steady state, so the policy flip can't dismiss the .transient
  popover (code-reviewer).
- Replace the non-cancellable asyncAfter auto-dismiss with a
  cancellable Task, and clear the retained popover via
  NSPopoverDelegate.popoverDidClose (covers the outside-click path).
- Split the wizard lifecycle into AppDelegate+Onboarding.swift and
  lift DiscoveryPopoverView to its own file: AppDelegate had
  crossed the 350-line hard ceiling (§2.1). Extract
  fireDiscoveryHintIfPending() so windowWillClose names its jobs.
