+++ docs/user-guide.md
@@ -75,10 +75,12 @@ logging out and back in.
 When setup finishes, a final card confirms KiwiDesk is now
 arranging your windows and offers to **Open Settings** (landing on
 Layout) — or **Not Now** to jump straight in; neither is required.
-A one-time hint then points at the menu bar icon, where KiwiDesk
-lives, so it's easy to find again. Both appear only once, on that
-first setup — losing and regranting Accessibility later never
-repeats them.
+If you choose Not Now (or just close the card), a one-time hint
+points at the menu bar icon, where KiwiDesk lives, so it's easy to
+find again; opening Settings skips the hint, since it already
+takes you into the app. Both appear only once, on that first
+setup — losing and regranting Accessibility later never repeats
+them.
 
 If Accessibility permission is ever missing — you dismissed the
 wizard, or the permission was revoked later — window management
